Fix db flow & add exceptions and handler (#3)

+++ src/main/java/ro/unibuc/hello/service/MovieService.java
@@ -11,8 +11,10 @@ import ro.unibuc.hello.dto.tmdb.CastDto;
 import ro.unibuc.hello.dto.tmdb.GenreDto;
 import ro.unibuc.hello.dto.tmdb.MovieApiDto;
 import ro.unibuc.hello.dto.tmdb.PagedApiResponseDto;
+import ro.unibuc.hello.exception.EntityNotFoundException;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -33,13 +35,12 @@ public class MovieService {
         return movies;
     }
 
-    private List<ActorDto> getActors(Long tmdbMovieId) {
-        final CastDto cast = theMovieDbClient.getCastByMovieId(tmdbMovieId);
-        return cast.getCast();
-    }
-
-    public Movie addMovie(Long tmbbId) {
-        final MovieApiDto movieApiDto = theMovieDbClient.getMovieById(tmbbId);
+    public Movie addMovie(Long tmdbId) {
+        final Optional<Movie> existingMovie = movieRepository.findByTmdbId(tmdbId);
+        if (existingMovie.isPresent()) {
+            throw new IllegalArgumentException("Movie already exists");
+        }
+        final MovieApiDto movieApiDto = theMovieDbClient.getMovieById(tmdbId);
         final List<String> genres = movieApiDto.getGenres()
                 .stream()
                 .map(GenreDto::getName)
@@ -52,10 +53,11 @@ public class MovieService {
                 .genres(genres)
                 .year(releaseYear)
                 .popularity(movieApiDto.getPopularity())
+                .tmdbId(tmdbId)
                 .build();
 
         final Movie savedMovie = movieRepository.save(movie);
-        final List<ActorDto> cast = getActors(tmbbId);
+        final List<ActorDto> cast = getActors(tmdbId);
 
         cast.stream()
                 .map(this::getActorOrAdd)
@@ -64,20 +66,30 @@ public class MovieService {
         return savedMovie;
     }
 
-    private Actor getActorOrAdd(ActorDto actorDto) {
-        return actorService.getActorByTmdbId(actorDto.getTmdbId())
-                .orElseGet(() -> actorService.saveActor(actorDto));
-    }
-
-    public Movie getMovieById(Long id) {
-        return movieRepository.findById(id).orElseThrow();
+    public Movie getMovieById(String id) {
+        return movieRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException(String.format("movie (id = %s)", id)));
     }
 
     public List<Movie> getMovies() {
         return movieRepository.findAll();
     }
 
-    public void deleteMovie(Long id) {
+    public void deleteMovie(String id) {
+        if (!movieRepository.existsById(id)) {
+            throw new EntityNotFoundException(String.format("movie (id = %s)", id));
+        }
+        roleService.deleteRolesByMovieId(id);
         movieRepository.deleteById(id);
+    }
+
+    private Actor getActorOrAdd(ActorDto actorDto) {
+        return actorService.getActorByTmdbId(actorDto.getTmdbId())
+                .orElseGet(() -> actorService.saveActor(actorDto));
+    }
+
+    private List<ActorDto> getActors(Long tmdbMovieId) {
+        final CastDto cast = theMovieDbClient.getCastByMovieId(tmdbMovieId);
+        return cast.getCast();
     }
 }
